LayoutTests/fast/forms/datalist/datalist-textinput-keydown-expected.txt

Test for keyboard operations for <input type=text> with list attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "Apple"
PASS input.value is "Pear"
PASS successfullyParsed is true

TEST COMPLETE
